+++ app/assets/stylesheets/public/shops.scss
@@ -57,6 +57,13 @@ shop-image {
   display: inline-block;
 }
 
+.comment-image {
+  
+  // width: 300px;
+  // height: 300px;
+  // object-fit: cover;
+}
+
 shop-item {
   display: inline-block;
 }
