+++ app/helpers/public/search_helper.rb
@@ -1,0 +1,2 @@
+module Public::SearchHelper
+end
