+++ app/views/layouts/application.html.erb
@@ -19,6 +19,14 @@
         <li class="nav-item"><%= link_to "マイページ", customer_path(current_customer), class: 'link'%></li>
         <li class="nav-item"><%= link_to "店舗一覧", shops_path, class: 'link'%></li>
         <li class="nav-item"><%= link_to "ログアウト", destroy_customer_session_path, method: :delete, class: 'link' %></li>
+        <li class="nav-item">
+          <%= form_with url: shops_path, method: :get, local: true do |f| %>
+            <%= f.text_field 'search[value]' %>
+            <span class="search-button text-right">
+              <%= f.submit :"検索" %>
+            </span>
+          <% end %>
+        </li>
       </ul>
     </div>
     <% elsif admin_signed_in? %>
