+++ app/views/public/contacts/confirm.html.erb
@@ -5,20 +5,20 @@
       <td><%= @contact.name %></td>
     </tr>
     <tr>
-      <td class="text-center">メッセージ</td>
+      <td class="text-center">お問い合わせ内容</td>
       <td style="white-space: pre-wrap;"><%= @contact.message %></td>
     </tr>
   </tbody>
 </table>
 
-<%= form_for(@contact) do |f| %>
+<%= form_with model: @contact do |f| %>
   <%= f.hidden_field :name %>
   <%= f.hidden_field :email %>
   <%= f.hidden_field :title %>
   <%= f.hidden_field :message %>
   <div><%= f.submit '送信' %></div>
 <% end %>
-<%= form_for @contact, url: back_path do |f| %>
+<%= form_with model:  @contact, url: back_path do |f| %>
   <%= f.hidden_field :name %>
   <%= f.hidden_field :email %>
   <%= f.hidden_field :title %>
