+++ app/views/public/contacts/new.html.erb
@@ -1,15 +1,19 @@
-<%= form_with model: @contact, url: confirm_path do |f| %>
-  <div class="form-group">
-    <%= f.label :name, 'お名前*' %>
-    <%= f.text_field :name, autofocus: true, class: 'form-control' %>
+<div class="container">
+  <div class="row justify-content-center">
+    <div class="col-md-8">
+      <%= form_with model: @contact, url: confirm_path do |f| %>
+        <div class="form-group my-3">
+          <%= f.label :name, 'お名前*' %>
+          <%= f.text_field :name, autofocus: true, class: 'form-control' %>
+        </div>
+        <div class="form-group my-3">
+          <%= f.label :message, 'お問い合わせ内容*' %>
+          <%= f.text_area :message, size: '10x10', class: 'form-control' %>
+        </div>
+        <div>
+          <%= f.submit '入力内容確認' %>
+        </div>
+      <% end %>
+    </div>
   </div>
-
-  <div class="form-group">
-    <%= f.label :message, 'メッセージ*' %>
-    <%= f.text_area :message, size: '10x10', class: 'form-control' %>
-  </div>
-
-  <div>
-    <%= f.submit '入力内容確認' %>
-  </div>
-<% end %>
+</div>
