+++ app/views/public/search/search.html.erb
@@ -1,0 +1,78 @@
+<dev class="container">
+  <div class="row">
+    <div class="col-md-4">
+
+      <table class="table table-bordered genre-table">
+        <thead>
+          <tr>
+            <th>エリアから探す</th>
+          </tr>
+        </thead>
+        <tbody>
+        <% @shops.each do |shop| %>
+        <% area = shop.area %>
+          <tr>
+            <td>
+              <%= link_to area.name, shops_path('search[value]': area.id, 'search[how]': "match") %>
+            </td>
+          </tr>
+        <% end %>
+        </tbody>
+      </table>
+      
+      <table class="table table-bordered genre-table">
+        <thead>
+          <tr>
+            <th>雰囲気から探す</th>
+          </tr>
+        </thead>
+        <tbody>
+        <% @shops.each do |shop| %>
+        <% atmosphere = shop.atmosphere %>
+          <tr>
+            <td>
+              <%= link_to atmosphere.atmosphere, shops_path('search[value]': atmosphere.id, 'search[how]': "match") %>
+            </td>
+          </tr>
+        <% end %>
+        </tbody>
+      </table>
+
+    </div>
+    <div class="col-md-8">
+      <div class="row">
+        <h4>店舗一覧</h4>
+        <div>
+          <% @shops.each do |shop| %>
+          <div class="img-thumbnail">
+  			    <div class="col-md-5">
+  			      <div class="col-md-5 shop-image">
+                <%= link_to shop_path(shop), class: "btn-link" do %>
+                  <%= attachment_image_tag shop, :image, format: 'jpeg', size: "320x220" %>
+                  <%# カラムはimage_idだがrefileの仕様でidは書かない %>
+                <% end %>
+              </div>
+            </div>
+            <div class="col-md-7 shop-item">
+              <div id="favorite_shop_<%= shop.id %>">
+                <%= render 'favorite', shop: shop %>
+              </div>
+              <div id="candidate_shop_<%= shop.id %>">
+                <%= render 'candidate', shop: shop %>
+              </div>
+              <h5><span>店名  <%= shop.name %></span></h5>
+              <p>営業日  <%= shop.business_day %></p>
+              <p>販売形式<a>  </a>
+                <%= shop.eat_in == "1" ? "イートイン  " : "" %><%= shop.take_out == "1" ? "テイクアウト" : "" %>
+              </p>
+              <p>予算  <%= shop.budget %></p>
+              <p>エリア  <%= shop.area.name %></p>
+              <p>店舗ジャンル  <%= shop.atmosphere.atmosphere %></p>
+            </div>
+            </div>
+          <% end %>
+        </div>
+      </div>
+    </div>
+  </div>
+</dev>
