+++ app/views/public/shops/index.html.erb
@@ -2,7 +2,7 @@
   <div class="row">
     <div class="col-md-4">
 
-      <table class="table table-bordered genre-table">
+      <table class="table table-bordered genre-table" width="60">
         <thead>
           <tr>
             <th>エリアから探す</th>
@@ -11,6 +11,7 @@
         <tbody>
         <% @shops.each do |shop| %>
         <% area = shop.area %>
+        <%# @areas.each do |area| %>
           <tr>
             <td>
               <%= link_to area.name, shops_path('search[value]': area.id, 'search[how]': "match") %>
@@ -40,38 +41,53 @@
 
     </div>
     <div class="col-md-8">
-      <div class="row">
-        <h4>店舗一覧</h4>
-        <div>
-          <% @shops.each do |shop| %>
-          <div class="img-thumbnail">
-  			    <div class="col-md-5">
-  			      <div class="col-md-5 shop-image">
-                <%= link_to shop_path(shop), class: "btn-link" do %>
-                  <%= attachment_image_tag shop, :image, format: 'jpeg', size: "320x220" %>
-                  <%# カラムはimage_idだがrefileの仕様でidは書かない %>
-                <% end %>
+      <div class="table">
+        <thead>
+          <tr>
+            <th><h4>店舗一覧</h4></th>
+          </tr>
+        </thead>
+        
+        <tbody>
+          <tr>
+            <% @shops.each do |shop| %>
+              <div class="">
+                <td>
+                  <div class="shop-image">
+                    <%= link_to shop_path(shop), class: "btn-link" do %>
+                      <div class="img-thumbnail">
+                        <%= attachment_image_tag shop, :image, format: 'jpeg', size: "320x220" %>
+                      </div>
+                      <%# カラムはimage_idだがrefileの仕様でidは書かない %>
+                    <% end %>
+                  </div>
+    			      </td>
+    			      <td>
+                  <div class="shop-item">
+                    <div id="favorite_shop_<%= shop.id %>">
+                      <%= render 'favorite', shop: shop %>
+                    </div>
+                  </div>
+                </td>
+                <td>
+                  <div id="candidate_shop_<%= shop.id %>">
+                    <%= render 'candidate', shop: shop %>
+                  </div>
+                </td><br>
+                <td><h5><span>店名  <%= shop.name %></span></h5></td><br>
+                <td>営業日  <%= shop.business_day %></td><br>
+                <td>
+                  販売形式<a>  </a>
+                    <%= shop.eat_in == "1" ? "イートイン  " : "" %><%= shop.take_out == "1" ? "テイクアウト" : "" %>
+                </td><br>
+                <td>予算  <%= shop.budget %></td><br>
+                <td>エリア  <%= shop.area.name %></td><br>
+                <td>店舗ジャンル  <%= shop.atmosphere.atmosphere %></td><br>
               </div>
+            <% end %>
             </div>
-            <div class="col-md-7 shop-item">
-              <div id="favorite_shop_<%= shop.id %>">
-                <%= render 'favorite', shop: shop %>
-              </div>
-              <div id="candidate_shop_<%= shop.id %>">
-                <%= render 'candidate', shop: shop %>
-              </div>
-              <h5><span>店名  <%= shop.name %></span></h5>
-              <p>営業日  <%= shop.business_day %></p>
-              <p>販売形式<a>  </a>
-                <%= shop.eat_in == "1" ? "イートイン  " : "" %><%= shop.take_out == "1" ? "テイクアウト" : "" %>
-              </p>
-              <p>予算  <%= shop.budget %></p>
-              <p>エリア  <%= shop.area.name %></p>
-              <p>店舗ジャンル  <%= shop.atmosphere.atmosphere %></p>
-            </div>
-            </div>
-          <% end %>
-        </div>
+          </tr>
+        </tbody>
       </div>
     </div>
   </div>
