+++ app/views/public/shops/show.html.erb
@@ -61,9 +61,13 @@
 
     <div class="tabbox-contents">
       <div class="tabbox box-show">
+        
         <% @shop.comments.each do |comment| %>
-          <p><%= attachment_image_tag(comment, :image, size: "350x230") %></p>
-          <p><%# if comment.customer == current_user %><%#= link_to "削除", shop_comment_path(comment.shop_id, comment.id), method: :delete %><%# end %></p>
+        <div class="comment-image">
+          <div class="col-mb-3">
+          <p><%= attachment_image_tag(comment, :image, size: "300x300") %></p>
+      　</div>
+      　</div>
         <% end %>
       </div>
