+++ test/controllers/public/search_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Public::SearchControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
